requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-api-tracking==1.8.4
flake8==6.1.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
kombu==5.3.4
mccabe==0.7.0
outcome==1.3.0
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.41
psycopg2==2.9.8
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.11
whitenoise==6.5.0
wsproto==1.2.0
